moving CLI to cmd/ (#362)
* moving CLI to cmd/
