+++ .gitignore
@@ -10,4 +10,4 @@ test/full/full
 gh-md-toc
 web/public/
 drand
-coverage.txt
+cover*
