+++ cmd/drand-cli/control.go
@@ -1,4 +1,4 @@
-package main
+package drand
 
 import (
 	"fmt"
@@ -21,7 +21,7 @@ func shareCmd(c *cli.Context) error {
 	var connectPeer net.Peer
 	if !isLeader {
 		if !c.IsSet(connectFlag.Name) {
-			fatal("need to the address of the coordinator to create the group file")
+			return fmt.Errorf("need to the address of the coordinator to create the group file")
 		}
 		coordAddress := c.String(connectFlag.Name)
 		isTls := !c.IsSet(insecureFlag.Name)
@@ -37,14 +37,14 @@ func shareCmd(c *cli.Context) error {
 		str := c.String(timeoutFlag.Name)
 		timeout, err = time.ParseDuration(str)
 		if err != nil {
-			fatal("dkg timeout duration incorrect:", err)
+			return fmt.Errorf("dkg timeout duration incorrect: %s", err)
 		}
 	}
 
 	conf := contextToConfig(c)
 	client, err := net.NewControlClient(conf.ControlPort())
 	if err != nil {
-		fatal("could not create client: %v", err)
+		return fmt.Errorf("could not create client: %v", err)
 	}
 
 	var groupP *control.GroupPacket
@@ -53,32 +53,35 @@ func shareCmd(c *cli.Context) error {
 		if c.IsSet(userEntropyOnlyFlag.Name) && !c.IsSet(sourceFlag.Name) {
 			fmt.Print("drand: userEntropyOnly needs to be used with the source flag, which is not specified here. userEntropyOnly flag is ignored.")
 		}
-		entropyInfo := entropyInfoFromReader(c)
+		entropyInfo, err := entropyInfoFromReader(c)
+		if err != nil {
+			return fmt.Errorf("error getting entropy source: %s", err)
+		}
 		if isLeader {
 			if !c.IsSet(periodFlag.Name) {
-				fatal("leader flag indicated requires the beacon period flag as well")
+				return fmt.Errorf("leader flag indicated requires the beacon period flag as well")
 			}
 			periodStr := c.String(periodFlag.Name)
 			period, err := time.ParseDuration(periodStr)
 			if err != nil {
-				fatal("period given is invalid: %v", err)
+				return fmt.Errorf("period given is invalid: %v", err)
 			}
 
 			offset := int(core.DefaultGenesisOffset.Seconds())
 			if c.IsSet(beaconOffset.Name) {
 				offset = c.Int(beaconOffset.Name)
 			}
-			fmt.Println("Initiating the DKG as a leader")
-			fmt.Println("You can stop the command at any point. If so, the group " +
-				"file will not be written out to the specified output. To get the" +
-				"group file once the setup phase is done, you can run the `drand show" +
+			fmt.Fprintln(output, "Initiating the DKG as a leader")
+			fmt.Fprintln(output, "You can stop the command at any point. If so, the group "+
+				"file will not be written out to the specified output. To get the"+
+				"group file once the setup phase is done, you can run the `drand show"+
 				"group` command")
 			groupP, shareErr = client.InitDKGLeader(nodes, thr, period, timeout, entropyInfo, secret, offset)
-			fmt.Println(" --- got err", shareErr, "group", groupP)
+			fmt.Fprintln(output, " --- got err", shareErr, "group", groupP)
 		} else {
-			fmt.Println("Participating to the setup of the DKG")
+			fmt.Fprintln(output, "Participating to the setup of the DKG")
 			groupP, shareErr = client.InitDKG(connectPeer, nodes, thr, timeout, entropyInfo, secret)
-			fmt.Println(" --- got err", shareErr, "group", groupP)
+			fmt.Fprintln(output, " --- got err", shareErr, "group", groupP)
 		}
 	} else {
 		// resharing case needs the previous group
@@ -89,7 +92,7 @@ func shareCmd(c *cli.Context) error {
 		} else if c.IsSet(oldGroupFlag.Name) {
 			var oldGroup = new(key.Group)
 			if err := key.Load(c.String(oldGroupFlag.Name), oldGroup); err != nil {
-				fatal("could not load drand from path", err)
+				return fmt.Errorf("could not load drand from path: %s", err)
 			}
 			oldPath = c.String(oldGroupFlag.Name)
 		}
@@ -99,97 +102,109 @@ func shareCmd(c *cli.Context) error {
 			if c.IsSet(beaconOffset.Name) {
 				offset = c.Int(beaconOffset.Name)
 			}
-			fmt.Println("Initiating the resharing as a leader")
+			fmt.Fprintln(output, "Initiating the resharing as a leader")
 			groupP, shareErr = client.InitReshareLeader(nodes, thr, timeout, secret, oldPath, offset)
 		} else {
-			fmt.Println("Participating to the resharing")
+			fmt.Fprintln(output, "Participating to the resharing")
 			groupP, shareErr = client.InitReshare(connectPeer, nodes, thr, timeout, secret, oldPath)
 		}
 	}
 	if shareErr != nil {
-		fatal("error setting up the network: %v", err)
+		return fmt.Errorf("error setting up the network: %v", err)
 	}
 	group, err := key.GroupFromProto(groupP)
 	if err != nil {
-		fatal("error interpreting the group from protobuf: %v", err)
+		return fmt.Errorf("error interpreting the group from protobuf: %v", err)
 	}
-	groupOut(c, group)
-	return nil
+	return groupOut(c, group)
 }
 
 func getShare(c *cli.Context) error {
-	client := controlClient(c)
+	client, err := controlClient(c)
+	if err != nil {
+		return err
+	}
 	resp, err := client.Share()
 	if err != nil {
-		fatal("drand: could not request the share: %s", err)
+		return fmt.Errorf("could not request the share: %s", err)
 	}
-	printJSON(resp)
-	return nil
+	return printJSON(resp)
 }
 
 func pingpongCmd(c *cli.Context) error {
-	client := controlClient(c)
-	if err := client.Ping(); err != nil {
-		fatal("drand: can't ping the daemon ... %s", err)
+	client, err := controlClient(c)
+	if err != nil {
+		return err
 	}
-	fmt.Printf("drand daemon is alive on port %s", controlPort(c))
+	if err := client.Ping(); err != nil {
+		return fmt.Errorf("drand: can't ping the daemon ... %s", err)
+	}
+	fmt.Fprintf(output, "drand daemon is alive on port %s", controlPort(c))
 	return nil
 }
 
 func showGroupCmd(c *cli.Context) error {
-	client := controlClient(c)
+	client, err := controlClient(c)
+	if err != nil {
+		return err
+	}
 	r, err := client.GroupFile()
 	if err != nil {
-		fatal("drand: fetching group file error: %s", err)
+		return fmt.Errorf("fetching group file error: %s", err)
 	}
 	group, err := key.GroupFromProto(r)
 	if err != nil {
 		return err
 	}
-	groupOut(c, group)
-	return nil
+	return groupOut(c, group)
 }
 
 func showCokeyCmd(c *cli.Context) error {
-	client := controlClient(c)
+	client, err := controlClient(c)
+	if err != nil {
+		return err
+	}
 	resp, err := client.CollectiveKey()
 	if err != nil {
-		fatal("drand: could not request drand.cokey: %s", err)
+		return fmt.Errorf("could not request drand.cokey: %s", err)
 	}
-	printJSON(resp)
-	return nil
+	return printJSON(resp)
 }
 
 func showPrivateCmd(c *cli.Context) error {
-	client := controlClient(c)
+	client, err := controlClient(c)
+	if err != nil {
+		return err
+	}
 	resp, err := client.PrivateKey()
 	if err != nil {
-		fatal("drand: could not request drand.private: %s", err)
+		return fmt.Errorf("could not request drand.private: %s", err)
 	}
-	printJSON(resp)
-	return nil
+	return printJSON(resp)
 }
 
 func showPublicCmd(c *cli.Context) error {
-	client := controlClient(c)
+	client, err := controlClient(c)
+	if err != nil {
+		return err
+	}
 	resp, err := client.PublicKey()
 	if err != nil {
-		fatal("drand: could not request drand.public: %s", err)
+		return fmt.Errorf("drand: could not request drand.public: %s", err)
 	}
-
-	printJSON(resp)
-	return nil
+	return printJSON(resp)
 }
 
 func showShareCmd(c *cli.Context) error {
-	client := controlClient(c)
+	client, err := controlClient(c)
+	if err != nil {
+		return err
+	}
 	resp, err := client.Share()
 	if err != nil {
-		fatal("drand: could not request drand.share: %s", err)
+		return fmt.Errorf("could not request drand.share: %s", err)
 	}
-
-	printJSON(resp)
-	return nil
+	return printJSON(resp)
 }
 
 func controlPort(c *cli.Context) string {
@@ -200,35 +215,36 @@ func controlPort(c *cli.Context) string {
 	return port
 }
 
-func controlClient(c *cli.Context) *net.ControlClient {
+func controlClient(c *cli.Context) (*net.ControlClient, error) {
 	port := controlPort(c)
 	client, err := net.NewControlClient(port)
 	if err != nil {
-		fatal("drand: can't instantiate control client: %s", err)
+		return nil, fmt.Errorf("can't instantiate control client: %s", err)
 	}
-	return client
+	return client, nil
 }
 
-func printJSON(j interface{}) {
+func printJSON(j interface{}) error {
 	buff, err := json.MarshalIndent(j, "", "    ")
 	if err != nil {
-		fatal("drand: could not JSON marshal: %s", err)
+		return fmt.Errorf("could not JSON marshal: %s", err)
 	}
-	fmt.Println(string(buff))
+	fmt.Fprintln(output, string(buff))
+	return nil
 }
 
-func entropyInfoFromReader(c *cli.Context) *control.EntropyInfo {
+func entropyInfoFromReader(c *cli.Context) (*control.EntropyInfo, error) {
 	if c.IsSet(sourceFlag.Name) {
 		_, err := os.Lstat(c.String(sourceFlag.Name))
 		if err != nil {
-			fatal("drand: cannot use given entropy source: %s", err)
+			return nil, fmt.Errorf("cannot use given entropy source: %s", err)
 		}
 		source := c.String(sourceFlag.Name)
 		ei := &control.EntropyInfo{
 			Script:   source,
 			UserOnly: c.Bool(userEntropyOnlyFlag.Name),
 		}
-		return ei
+		return ei, nil
 	}
-	return nil
+	return nil, nil
 }
