+++ cmd/drand-cli/public.go
@@ -1,7 +1,8 @@
-package main
+package drand
 
 import (
 	"errors"
+	"fmt"
 	gonet "net"
 
 	"github.com/drand/drand/core"
@@ -19,10 +20,12 @@ func getPrivateCmd(c *cli.Context) error {
 	if c.IsSet("tls-cert") {
 		defaultManager.Add(c.String("tls-cert"))
 	}
-	ids := getNodes(c)
+	ids, err := getNodes(c)
+	if err != nil {
+		return err
+	}
 	client := core.NewGrpcClientFromCert(defaultManager)
 	var resp []byte
-	var err error
 	for _, public := range ids {
 		resp, err = client.Private(public.Identity)
 		if err == nil {
@@ -33,7 +36,7 @@ func getPrivateCmd(c *cli.Context) error {
 		slog.Infof("drand: error contacting node %s: %s", public.Addr, err)
 	}
 	if resp == nil {
-		slog.Fatalf("drand: zero successful contacts with nodes")
+		return errors.New("zero successful contacts with nodes")
 	}
 
 	type private struct {
@@ -55,15 +58,20 @@ func getPublicRandomness(c *cli.Context) error {
 		client = core.NewGrpcClientFromCert(defaultManager)
 	}
 
-	ids := getNodes(c)
-	group := getGroup(c)
+	ids, err := getNodes(c)
+	if err != nil {
+		return err
+	}
+	group, err := getGroup(c)
+	if err != nil {
+		return err
+	}
 	if group.PublicKey == nil {
-		slog.Fatalf("drand: group file must contain the distributed public key!")
+		return errors.New("drand: group file must contain the distributed public key!")
 	}
 
 	public := group.PublicKey
 	var resp *drand.PublicRandResponse
-	var err error
 	var foundCorrect bool
 	for _, id := range ids {
 		if c.IsSet(roundFlag.Name) {
@@ -98,7 +106,7 @@ func getCokeyCmd(c *cli.Context) error {
 	for _, addr := range c.Args().Slice() {
 		_, _, err := gonet.SplitHostPort(addr)
 		if err != nil {
-			fatal("invalid address given: %s", err)
+			return fmt.Errorf("invalid address given: %s", err)
 		}
 		dkey, err = client.DistKey(addr, !c.Bool("tls-disable"))
 		if err == nil {
@@ -108,8 +116,7 @@ func getCokeyCmd(c *cli.Context) error {
 			addr, err)
 	}
 	if dkey == nil {
-		slog.Fatalf("drand: can't retrieve dist. key from all nodes")
+		return errors.New("can't retrieve dist. key from all nodes")
 	}
-	printJSON(dkey)
-	return nil
+	return printJSON(dkey)
 }
